requirements.txt
imapclient
pyzmail36
requests
python-dotenv
selenium
webdriver-manager
google-genai